requirements.txt
Django==3.1.2
djangorestframework==3.11.0
django-filter==2.4.0
drf-yasg==1.17.1
PyJWT==1.6.4
gunicorn==20.0.4
uvicorn==0.11.5
django-safedelete==0.5.5
firebase-admin==4.3.0
python-slugify==4.0.1
django-storages==1.9.1
boto3==1.12.6
python-dateutil==2.8.1
requests==2.23.0
psycopg2==2.8.4